requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
